+++ tests/configuration/tests.py
@@ -341,7 +341,7 @@ class TestLoadConfigDecorator(TestCase):
         })
 
     @staticmethod
-    @load_config(defaults=MOCK_CONFIGURATION)
+    @load_config()
     def decorated(config, *args, **kwargs):
         return config, args, kwargs
 
@@ -353,7 +353,11 @@ class TestLoadConfigDecorator(TestCase):
             config=self.config.to_dict(protected=True, private=True)
         )
         self.assertIsInstance(test_config, ConfigurationType)
-        self.assertEqual(test_config._defaults, MOCK_CONFIGURATION)
+        self.assertIsNone(
+            test_config._defaults,
+            "Expected load_config to initialize without default. "
+            "Use register_defaults to set defaults for load_config decorator."
+        )
         self.assertEqual(test_config._namespace, "name")
         self.assertIn("_test3", test_config._private)
         self.assertEqual(self.config.test, "public")