+++ CHANGELOG.md
@@ -62,6 +62,7 @@ The convenience of lists is more important here than memory footprint which will
 * For consistent ``Resource`` serialization adds ``serialize_resources`` and ``update_serialized_resources``.
 * Experimental support for ``ResourceFixturesMixin`` that can be used to load resource content through fixture files.
 * Cancelling a ``HttpFileResource`` will result in an empty body instead of a body of None.
+* Removes ``load_config`` defaults parameter.
 
 
 v0.18